Move picojson out to its own file
